pyproject.toml
[project]
name = "friends-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.22.0",
    "pydantic>=2.11.10",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
]
